WIP to implement syntax to "hide" tyvars from visible application
This commit does not currently work properly, but it compiles.

The following file compiles properly with this commit of GHC,
but this is *not* the expected behavior:

```
foo :: forall {a :: *} {b :: *}. b -> (b, a)
foo x = (x, undefined)

bar :: (Bool, Int)
bar = foo @Int True
```

The visible type application above should not have been allowed
at all, because both `a` and `b` are set as "inferred" by the
curly-brace syntax. Instead, we are allowed to apply `foo` to
the type Int, as if nothing had changed.

In particular, there are some remaining issues:

- The parser now has 57 shift-reduce conflicts, up from 36. I'm not
  certain how to fix this, but it's certainly broken somehow.

- It appears that even though the ArgFlag information now gets sent
  throughout all the parts of the front-end which needed to be
  refactored to accommodate it, it still doesn't result in the
  desired behavior. Visible type application still works on
  arguments which should have been hidden by curly-brace syntax.

- In various places throughout the compiler, there are functions
  which try to manufacture a new HsTyVarBndr ex nihilo, without
  a starting ArgFlag to determine what sort of binder they need
  to produce. I've tried to guess, and noted where I've made such
  guesses with TODO notes (grep the diff for this commit for TODO
  to locate these places). I'm not so sure about a lot of these
  locations, and I'm not sure in general how to deal with this
  kind of issue.